refactor: Implement `Main Menu` orcherstrator
